+++ src/event.rs
@@ -13,12 +13,12 @@
 // TODO: I'm not sure that these events are/can be safe by Rust's model of safety; they inherently
 // create state which can be mutated even while an immutable borrow is held.
 
-use crate::cuda_sys::cuda::{
+use crate::error::{CudaError, CudaResult, DropResult, ToResult};
+use crate::stream::Stream;
+use cuda_driver_sys::{
     cuEventCreate, cuEventDestroy_v2, cuEventElapsedTime, cuEventQuery, cuEventRecord,
     cuEventSynchronize, CUevent,
 };
-use crate::error::{CudaError, CudaResult, DropResult, ToResult};
-use crate::stream::Stream;
 
 use std::mem;
 use std::ptr;