+++ src/stream.rs
@@ -13,7 +13,7 @@
 use crate::error::{CudaResult, DropResult, ToResult};
 use crate::event::Event;
 use crate::function::{BlockSize, Function, GridSize};
-use cuda_sys::cuda::{self, cudaError_t, CUstream};
+use cuda_driver_sys::{cudaError_enum, CUstream};
 use std::ffi::c_void;
 use std::mem;
 use std::panic;
@@ -88,7 +88,7 @@ impl Stream {
             let mut stream = Stream {
                 inner: ptr::null_mut(),
             };
-            cuda::cuStreamCreateWithPriority(
+            cuda_driver_sys::cuStreamCreateWithPriority(
                 &mut stream.inner as *mut CUstream,
                 flags.bits(),
                 priority.unwrap_or(0),
@@ -117,7 +117,7 @@ impl Stream {
     pub fn get_flags(&self) -> CudaResult<StreamFlags> {
         unsafe {
             let mut bits = 0u32;
-            cuda::cuStreamGetFlags(self.inner, &mut bits as *mut u32).to_result()?;
+            cuda_driver_sys::cuStreamGetFlags(self.inner, &mut bits as *mut u32).to_result()?;
             Ok(StreamFlags::from_bits_truncate(bits))
         }
     }
@@ -145,7 +145,8 @@ impl Stream {
     pub fn get_priority(&self) -> CudaResult<i32> {
         unsafe {
             let mut priority = 0i32;
-            cuda::cuStreamGetPriority(self.inner, &mut priority as *mut i32).to_result()?;
+            cuda_driver_sys::cuStreamGetPriority(self.inner, &mut priority as *mut i32)
+                .to_result()?;
             Ok(priority)
         }
     }
@@ -186,7 +187,7 @@ impl Stream {
         T: FnOnce(CudaResult<()>) + Send,
     {
         unsafe {
-            cuda::cuStreamAddCallback(
+            cuda_driver_sys::cuStreamAddCallback(
                 self.inner,
                 Some(callback_wrapper::<T>),
                 Box::into_raw(callback) as *mut c_void,
@@ -219,7 +220,7 @@ impl Stream {
     /// # }
     /// ```
     pub fn synchronize(&self) -> CudaResult<()> {
-        unsafe { cuda::cuStreamSynchronize(self.inner).to_result() }
+        unsafe { cuda_driver_sys::cuStreamSynchronize(self.inner).to_result() }
     }
 
     /// Make the stream wait on an event.
@@ -253,7 +254,10 @@ impl Stream {
     /// }
     /// ```
     pub fn wait_event(&self, event: Event, flags: StreamWaitEventFlags) -> CudaResult<()> {
-        unsafe { cuda::cuStreamWaitEvent(self.inner, event.as_inner(), flags.bits()).to_result() }
+        unsafe {
+            cuda_driver_sys::cuStreamWaitEvent(self.inner, event.as_inner(), flags.bits())
+                .to_result()
+        }
     }
 
     // Hidden implementation detail function. Highly unsafe. Use the `launch!` macro instead.
@@ -273,7 +277,7 @@ impl Stream {
         let grid_size: GridSize = grid_size.into();
         let block_size: BlockSize = block_size.into();
 
-        cuda::cuLaunchKernel(
+        cuda_driver_sys::cuLaunchKernel(
             func.to_inner(),
             grid_size.x,
             grid_size.y,
@@ -329,7 +333,7 @@ impl Stream {
 
         unsafe {
             let inner = mem::replace(&mut stream.inner, ptr::null_mut());
-            match cuda::cuStreamDestroy_v2(inner).to_result() {
+            match cuda_driver_sys::cuStreamDestroy_v2(inner).to_result() {
                 Ok(()) => {
                     mem::forget(stream);
                     Ok(())
@@ -348,7 +352,7 @@ impl Drop for Stream {
         unsafe {
             let inner = mem::replace(&mut self.inner, ptr::null_mut());
             // No choice but to panic here.
-            cuda::cuStreamDestroy_v2(inner)
+            cuda_driver_sys::cuStreamDestroy_v2(inner)
                 .to_result()
                 .expect("Failed to destroy CUDA stream.");
         }
@@ -356,7 +360,7 @@ impl Drop for Stream {
 }
 unsafe extern "C" fn callback_wrapper<T>(
     _stream: CUstream,
-    status: cudaError_t,
+    status: cudaError_enum,
     callback: *mut c_void,
 ) where
     T: FnOnce(CudaResult<()>) + Send,